WideChar function use __int64

+++ phpwb.h
@@ -71,10 +71,10 @@ zval *process_array(zval *zitems);
 
 // String encode converting function
 
-TCHAR *Utf82WideChar(const char *str, int len);
-void Utf82WideCharCopy(const char *str, int str_len, TCHAR *wcs, int wcs_len);
+TCHAR *Utf82WideChar(const char *str, __int64 len);
+void Utf82WideCharCopy(const char *str, __int64 str_len, TCHAR *wcs, __int64 wcs_len);
 char *WideChar2Utf8(LPCTSTR wcs, __int64 *len);
-void WideCharCopy(LPCTSTR wcs, char *s, int len);
+void WideCharCopy(LPCTSTR wcs, char *s, __int64 len);
 void dumptcs(TCHAR *str);
 char *ConvertUTF16ToUTF8(LPCWSTR pszTextUTF16, int *plen);
